Trying to implement the Eight Steps Evaluation Scheme

+++ src/data-structures/Node.hpp
@@ -17,8 +17,8 @@ public:
   int id;
   int load;
   Point *point;
-  int serviceTime;
-  int maxRideTime;
+  float serviceTime;
+  float maxRideTime;
   float departureTime;
   float arrivalTime;
   float rechargingRate;
